services/py/payment/app/routes/payments.py
from typing import Annotated
from uuid import UUID

import jwt
from fastapi import APIRouter, Depends, Header, Query

from common.errors import AppError
from app.domain.payment import (
    PaymentCreate,
    PaymentResponse,
    PaymentListResponse,
)
from app.services.payment_service import PaymentService

router = APIRouter(prefix="/payments", tags=["payments"])


def _get_payment_service() -> PaymentService:
    from app.main import get_payment_service
    return get_payment_service()


def _get_current_user_claims(authorization: Annotated[str, Header()]) -> dict:
    from app.main import app_settings

    if not authorization.startswith("Bearer "):
        raise AppError("Invalid authorization header", status_code=401)
    token = authorization[7:]
    try:
        payload = jwt.decode(
            token, app_settings.jwt_secret, algorithms=[app_settings.jwt_algorithm]
        )
        return {
            "user_id": UUID(payload["sub"]),
            "role": payload.get("role", "student"),
            "is_verified": payload.get("is_verified", False),
        }
    except (jwt.PyJWTError, ValueError, KeyError) as exc:
        raise AppError("Invalid token", status_code=401) from exc


def _to_response(p: "Payment") -> PaymentResponse:
    from app.domain.payment import Payment
    return PaymentResponse(
        id=p.id,
        student_id=p.student_id,
        course_id=p.course_id,
        amount=p.amount,
        status=p.status,
        created_at=p.created_at,
    )


@router.post("", response_model=PaymentResponse, status_code=201)
async def create_payment(
    body: PaymentCreate,
    claims: Annotated[dict, Depends(_get_current_user_claims)],
    service: Annotated[PaymentService, Depends(_get_payment_service)],
) -> PaymentResponse:
    payment = await service.create(
        student_id=claims["user_id"],
        role=claims["role"],
        course_id=body.course_id,
        amount=body.amount,
    )
    return _to_response(payment)


@router.get("/me", response_model=PaymentListResponse)
async def list_my_payments(
    claims: Annotated[dict, Depends(_get_current_user_claims)],
    service: Annotated[PaymentService, Depends(_get_payment_service)],
    limit: Annotated[int, Query(ge=1, le=100)] = 20,
    offset: Annotated[int, Query(ge=0)] = 0,
) -> PaymentListResponse:
    items, total = await service.list_my(claims["user_id"], limit, offset)
    return PaymentListResponse(
        items=[_to_response(p) for p in items],
        total=total,
    )


@router.get("/{payment_id}", response_model=PaymentResponse)
async def get_payment(
    payment_id: UUID,
    claims: Annotated[dict, Depends(_get_current_user_claims)],
    service: Annotated[PaymentService, Depends(_get_payment_service)],
) -> PaymentResponse:
    p = await service.get(payment_id)
    return _to_response(p)
