services/py/payment/pyproject.toml
[project]
name = "payment"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "asyncpg>=0.30",
    "pyjwt>=2.10",
    "pydantic-settings>=2.7",
    "prometheus-fastapi-instrumentator>=7.0",
    "common",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25",
    "httpx>=0.28",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.uv.sources]
common = { workspace = true }
